Added logging; separated pico related files print() statements were causing an eventual hang when no USB console was connected so logging was added with print() called conditionally if a USB console is connected.
Logging always writes to a in-memory buffer so that messages can be
read using bluetooth serial console or re-connected USB console.

All pico related files were moved to the picow directory so that all
files in the picow directory can be copied to the board using the
rshell rsync command (and files no long need to be renamed).

Code for measuring temperature and power have been moved to separate
files.